requirements.txt
Flask
python-dotenv
azure-ai-inference
azure-core